Changed QueryRowContext to ExecContext in insert, update and delete o…

+++ entity/sentence.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Sentence struct {
-	SentenceID   int            `json:"sentence_id"`
+	SentenceID   int64          `json:"sentence_id"`
 	Body         string         `json:"body"`
 	Vocabularies pq.StringArray `json:"vocabularies"`
 	Created      string         `json:"created"`
